neck controll and motion manager change neck motors format

+++ src/control/src/control.cpp
@@ -104,7 +104,7 @@ public:
 
     }
 
-  vector<int> motors { 1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20 };
+  vector<u_int8_t> motors { 1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20 };
   vector<vector<int>> position;
   
 private:
